How to Make Responsive Animated Form (Multi Step)
In this, I'm going to show you, how to design a Multi Step Login and Signup Form using HTML CSS and a few line of JavaScript Code. It is very easy to design. Here I used eye button to show/hide password. Hope it will be helpful for the beginners.